Re-add suppression, add suppression for suppression
Yo dawg, I heard you like like suppressions, so we put a suppression
on a suppression, so it can suppress while suppressing...

cppcheck on OS X seems to think, that the autoVariables suppression
is unmatched. Removing it works fine. Under OS X and windows the same
thing does not work, it yields errors. So a suppressing that error
of the original suppression helps.